[NVIDIA] Reuse grouped MM test files for W8A8 FP8
