+++ conf/ci_test_aliases.yaml
@@ -44,3 +44,4 @@ arctan: test_atan
 # Operators tested in combined files or with different naming
 batch_norm_backward: test_batch_norm
 batch_norm_impl_index: test_batch_norm
+group_gemm_w8a8_fp8: test_grouped_mm
